Data_collection_app/requirements.txt
Flask
sounddevice
soundfile
google-cloud-storage
google-cloud-secret-manager
gunicorn